Fixed huge bug, added -o switch.
There was a huge bug in the tags parsing where I left the
mootools-core.js hardcoded. It now appropriately uses input.name.

I added the -o (--objects) switch to allow users to pass in a list of
objects to export if they don't want to use jsctags and already know
what they're looking for.